Move Console/Command under Command/Command.
The Command class has dependency on ORM package and hence having it
under standalone console package doesn't seem right.
