+++ src/Console/Command.php
@@ -2,62 +2,10 @@
 declare(strict_types=1);
 
 /**
- * CakePHP(tm) : Rapid Development Framework (https://cakephp.org)
- * Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
- *
- * Licensed under The MIT License
- * For full copyright and license information, please see the LICENSE.txt
- * Redistributions of files must retain the above copyright notice.
- *
- * @copyright     Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
- * @link          https://cakephp.org CakePHP(tm) Project
- * @since         3.6.0
- * @license       https://opensource.org/licenses/mit-license.php MIT License
+ * @deprecated 4.0.0 Use Cake\Command\Command instead.
  */
-namespace Cake\Console;
 
-use Cake\Datasource\ModelAwareTrait;
-use Cake\Log\LogTrait;
-use Cake\ORM\Locator\LocatorAwareTrait;
-
-/**
- * Base class for commands using the full stack
- * CakePHP Framework.
- *
- * Includes traits that integrate logging
- * and ORM models to console commands.
- */
-class Command extends BaseCommand
-{
-    use LocatorAwareTrait;
-    use LogTrait;
-    use ModelAwareTrait;
-
-    /**
-     * Constructor
-     *
-     * By default CakePHP will construct command objects when
-     * building the CommandCollection for your application.
-     */
-    public function __construct()
-    {
-        $this->modelFactory('Table', function ($alias) {
-            return $this->getTableLocator()->get($alias);
-        });
-
-        if (isset($this->modelClass)) {
-            $this->loadModel();
-        }
-    }
-
-    /**
-     * Implement this method with your command's logic.
-     *
-     * @param \Cake\Console\Arguments $args The command arguments.
-     * @param \Cake\Console\ConsoleIo $io The console io
-     * @return int|null|void The exit code or null for success
-     */
-    public function execute(Arguments $args, ConsoleIo $io)
-    {
-    }
-}
+class_alias(
+    'Cake\Command\Command',
+    'Cake\Console\Command'
+);
