+++ src/Console/Command/HelpCommand.php
@@ -17,8 +17,8 @@ declare(strict_types=1);
 namespace Cake\Console\Command;
 
 use ArrayIterator;
+use Cake\Command\Command;
 use Cake\Console\Arguments;
-use Cake\Console\Command;
 use Cake\Console\CommandCollection;
 use Cake\Console\CommandCollectionAwareInterface;
 use Cake\Console\ConsoleIo;
