+++ tests/TestCase/Command/CompletionCommandTest.php
@@ -16,7 +16,7 @@ declare(strict_types=1);
  */
 namespace Cake\Test\TestCase\Command;
 
-use Cake\Console\Command;
+use Cake\Command\Command;
 use Cake\Core\Configure;
 use Cake\Routing\Router;
 use Cake\TestSuite\ConsoleIntegrationTestCase;
