+++ tests/test_app/Plugin/TestPlugin/src/Command/WidgetCommand.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace TestPlugin\Command;
 
+use Cake\Command\Command;
 use Cake\Console\Arguments;
-use Cake\Console\Command;
 use Cake\Console\ConsoleIo;
 
 class WidgetCommand extends Command
