+++ tests/test_app/TestApp/Command/AutoLoadModelCommand.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace TestApp\Command;
 
-use Cake\Console\Command;
+use Cake\Command\Command;
 
 class AutoLoadModelCommand extends Command
 {
