+++ tests/test_app/TestApp/Command/DemoCommand.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace TestApp\Command;
 
+use Cake\Command\Command;
 use Cake\Console\Arguments;
-use Cake\Console\Command;
 use Cake\Console\ConsoleIo;
 
 class DemoCommand extends Command
